+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## next
 
 - added the `devices` capability to share all the devices and support hot-pluggable devices.
+- fix support for direct nixGL, see the readest app.
+- added readest.
 
 ## 0.7.0 (2025-02-14)