+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added the `devices` capability to share all the devices and support hot-pluggable devices.
 - fix support for direct nixGL, see the readest app.
 - added readest and freecad app.
+- added support for nix package names not matching their executable filename.
 
 ## 0.7.0 (2025-02-14)